Add concurrent iterative processing for file sync / file fold
We utilize iterative extent retrievals so we don't have to wait for all
extents to be retrieved, we also don't need to buffer the whole file extents
Since extents retrieval is a sys call, which moves control to the kernel
With an iterative retrieval we can ensure tha the process can be killed
instead of having to wait for all extents to be retrieved before control
returns to the application, we also limit the maximum consumed memory.

Longhorn #1948
Longhorn #2507

Add GetFileLayout function that retrieves the file intervals iteratively

The function returns a buffered data, err channel for processing of the
already determined intervals. The reason we use a buffered channel instead
of an unbounded queue, is that the buffered channel allows us to control
the maximum amount of allocated memory for the intervals/extents.

Implement Parallel Segment Processing

This change improves the processing speed of a 1GB file consisting of
4K data / 4K holes. On the local unit test, from 120s down to 9s

The block size of 4K was used, since it's the worst case scenario for
io latency / http overhead. Since each block requires a single request.

Example file sync with removed checksums check below:
segments: 262144/9.43s data: 131072/6.55s holes: 131072/2.83s

Update the temp interval buffer allocation size

This is so we only need at most a single allocation when converting extents
to intervals for processing. A little bit better design is to pre allocate
a shared buffer since each extent retrieval is 1024, we should only need
a single buffer of size = 1 + MaxExtentsBuffer * 2 then we can use slices
and tracking of the current buffer position to fill end consume that buffer.

That would make the code a little bit more complex, i.e more tracking.
So I decided for the simpler approach of 1 allocation per extent retrieval.
Since we run this as a separate process on demand.

Allow only a single inflight file fold progress update

We limit the update calls to p=[0,100] since there is no point in calling
update() for the same p value multiple times. Progress updates can only be
monotonically increasing, once that's replaced we can use an atomic cmp&swp
around progress that doesn't require locking.

Signed-off-by: Joshua Moody <joshua.moody@suse.com>